// src/pages/Products/styles.d.ts
import { ViewStyle, TextStyle } from 'react-native';

export default interface IProductsStyles {
  btn: ViewStyle,
  barBordered: ViewStyle,
  totalBar: ViewStyle,
  totalBarLabel: ViewStyle,
  totalBarValue: ViewStyle,
  totalBarText: TextStyle,
  listsBar: ViewStyle,
  listsBarPickerWrap: ViewStyle,
  listsBarPicker: ViewStyle,
  listsBarBtn: ViewStyle,
  listsModal:ViewStyle,
  listsModalText:TextStyle,
  listsModalInput:TextStyle,
  listsMenu:ViewStyle,
  listsMenuMask:ViewStyle,
  listsMenuArrow:ViewStyle,
  listsMenuItem:ViewStyle,
  listsMenuItemBordered:ViewStyle,
  products: ViewStyle,
  product: ViewStyle,
  productFieldSeparator: TextStyle,
  productBordered: TextStyle,
  productInput: TextStyle,
  productQtd: TextStyle,
  productValue: TextStyle,
  productBtn: TextStyle,
  productBtnBar: ViewStyle
}